# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Schedule Adjuster API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/user/{user_id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get User",
        "operationId": "get_user_user__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update User",
        "operationId": "update_user_user__user_id__put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Add User",
        "operationId": "add_user_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/schedules": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "Get Schedules",
        "operationId": "get_schedules_schedules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleGet"
                  },
                  "type": "array",
                  "title": "Response Get Schedules Schedules Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "schedule"
        ],
        "summary": "Update Schedules",
        "operationId": "update_schedules_schedules_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/ScheduleCreate"
                },
                "type": "array",
                "title": "Schedules"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleCreate"
                  },
                  "type": "array",
                  "title": "Response Update Schedules Schedules Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "Add Schedules",
        "operationId": "add_schedules_schedules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/ScheduleCreate"
                },
                "type": "array",
                "title": "Schedules"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduleCreate"
                  },
                  "type": "array",
                  "title": "Response Add Schedules Schedules Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/schedule/{schedule_id}": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "Get Schedule",
        "operationId": "get_schedule_schedule__schedule_id__get",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleGet"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "schedule"
        ],
        "summary": "Update Schedule",
        "operationId": "update_schedule_schedule__schedule_id__put",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleCreate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/schedule": {
      "post": {
        "tags": [
          "schedule"
        ],
        "summary": "Add Schedule",
        "operationId": "add_schedule_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleCreate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScheduleConfigGet": {
        "properties": {
          "schedule_master_id": {
            "type": "string",
            "title": "Schedule Master Id"
          },
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/ScheduleConstraintGet"
            },
            "type": "array",
            "title": "Constraints"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "schedule_master_id",
          "constraints",
          "created_at"
        ],
        "title": "ScheduleConfigGet"
      },
      "ScheduleConstraintGet": {
        "properties": {
          "schedule_config_id": {
            "type": "string",
            "title": "Schedule Config Id"
          },
          "prop_id": {
            "$ref": "#/components/schemas/ScheduleConstraintPropId"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "constraints": {
            "type": "object",
            "title": "Constraints"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "schedule_config_id",
          "prop_id",
          "name",
          "constraints",
          "created_at"
        ],
        "title": "ScheduleConstraintGet"
      },
      "ScheduleConstraintPropId": {
        "type": "string",
        "enum": [
          "intProp1",
          "intProp2",
          "intProp3",
          "listProp1",
          "listProp2",
          "listProp3"
        ],
        "title": "ScheduleConstraintPropId"
      },
      "ScheduleCreate": {
        "properties": {
          "schedule_master_id": {
            "type": "string",
            "title": "Schedule Master Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "title": "Startat"
          },
          "endAt": {
            "title": "Endat",
            "type": "string",
            "format": "date-time"
          },
          "allDay": {
            "type": "boolean",
            "title": "Allday"
          },
          "canceled": {
            "type": "boolean",
            "title": "Canceled"
          }
        },
        "type": "object",
        "required": [
          "schedule_master_id",
          "name",
          "startAt",
          "allDay",
          "canceled"
        ],
        "title": "ScheduleCreate"
      },
      "ScheduleGet": {
        "properties": {
          "schedule_master_id": {
            "type": "string",
            "title": "Schedule Master Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "title": "Startat"
          },
          "endAt": {
            "title": "Endat",
            "type": "string",
            "format": "date-time"
          },
          "allDay": {
            "type": "boolean",
            "title": "Allday"
          },
          "canceled": {
            "type": "boolean",
            "title": "Canceled"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "config": {
            "$ref": "#/components/schemas/ScheduleConfigGet"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "schedule_master_id",
          "name",
          "startAt",
          "allDay",
          "canceled",
          "id",
          "config",
          "created_at"
        ],
        "title": "ScheduleGet"
      },
      "UserCreate": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "birthday": {
            "title": "Birthday",
            "type": "string",
            "format": "date"
          }
        },
        "type": "object",
        "required": [
          "role",
          "name",
          "email"
        ],
        "title": "UserCreate"
      },
      "UserGet": {
        "properties": {
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "birthday": {
            "title": "Birthday",
            "type": "string",
            "format": "date"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "role",
          "name",
          "email",
          "id",
          "created_at"
        ],
        "title": "UserGet"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}